requirements/base.txt
Django==1.10.5
djangorestframework==3.5.3
django-webpack-loader==0.4.1
psycopg2==2.6.2